Update endpoint.rb
If the raw_post is stipulated don't calculate the params.  
I did this because when I want the default posts to be JSON, setting the config makes sense.  When I set the config and want to override it (for attachments) then I need the formatting bit of code to not be called because converting the attachment to JSON didn't work.