Removes utf-8 from JSON requests because it's redundant
JSON requests should use UTF-8 by default according to
http://www.ietf.org/rfc/rfc4627.txt, so we will remove `charset=utf-8`
when we find it to avoid redundancy.

In the process, I moved code that was only used by APIB into the APIB
classes, such as exposing `content_type` to the templates. If I changed
the `content-type` for all templates it would break unrelated things.

Connects to #235.