Alter example spec section
This has been altered to take out any open_api code. Given there
is only one example spec in the README document, my view is that it should
not be specifically open_api, but just use the most general options.
I have bumped the level of this section up by two levels, as it no longer fits
under the open_api section of the format section.

I have also taken out or re-factored instances of <response>, given
I could not understand what the code did in this instance.

I note that the example code in the repository is open_api, so there
is already an example of open_api code.