Remove dev dependency to Thin
This change is part of preliminary efforts at attempting to refresh
the Continuous Integration setup, by supporting recent Ruby version and
updating dependencies.

To be honest, I don't know what benefits the Thin server brings to the table.
I tried digging in the code, but the commit that introduced it does not
tell me much: cdeea8b.

What I know however is that I have troubles running the specs locally,
getting a timeout on this line:

https://github.com/zipmark/rspec_api_documentation/blob/560c3bdc7bd5581e7c223334390221ecfc910be8/spec/http_test_client_spec.rb#L16

However, letting Capybara handle its server details, as shown in this PR,
does not seem to have any negative impact (spec is still green, and fairly fast).