Test more recent versions of Ruby on Travis
Added:
- 2.4.9 (reached end-of-life but may temporarily help in incremental debugging)
- 2.5.8
- 2.6.6

Did not add 2.7.x yet because it introduces *a lot* of warnings.

Did not remove any past version yet because until decided otherwise,
they should be supported by the gem. (I would recommend a major version
bump when those versions get dropped.)

The idea is that developments from now on should be future-proof, and
changes should be tested on current versions of Ruby.